// src/model/Types.tsx
export type PaymentDetail = {
  amountToPay: number
  days: number
  hourSalary: number
  hoursPerDay: number
  laborCost: number
  matchingFee: number
  restTime: number
  salary: number
  transctionPath: string
  usedCash: number
  refundRate?: number
  isAllRefund?: boolean
  transactionPath: string
}